gemini/requirements.txt
streamlit_chat==0.1.1
streamlit==1.29.0
streamlit-chat==0.1.1
google-generativeai==0.3.1
python-dotenv==1.0.0